Python Scripts/tutorial/tutorial/spiders/requirements.txt
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
scrapy==2.4.1
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-cloud-logging==2.2.0
pymysql==1.0.2
sqlalchemy==1.3.21